internal/unionstore/art/art_test.go
package art

import (
	"testing"

	"github.com/stretchr/testify/assert"
	tikverr "github.com/tikv/client-go/v2/error"
)

func TestSimple(t *testing.T) {
	tree := New()
	for i := 0; i < 256; i++ {
		// key := []byte(strconv.Itoa(i))
		key := []byte{byte(i)}
		_, err := tree.Get(key)
		assert.Equal(t, err, tikverr.ErrNotExist)
		err = tree.Set(key, key)
		assert.Nil(t, err)
		val, err := tree.Get(key)
		assert.Nil(t, err, i)
		assert.Equal(t, val, key, i)
	}
}

// func BenchmarkTraverse(b *testing.B) {
// 	buf := make([][]byte, b.N)
// 	for i := range buf {
// 		buf[i] = []byte(strconv.Itoa(i))
// 	}

// 	tree := New()
// 	b.ResetTimer()
// 	for i := 0; i < b.N; i++ {
// 		tree.traverse(buf[i], true)
// 	}
// }

func BenchmarkReadAfterWriteArt(b *testing.B) {
	buf := make([][]byte, b.N)
	for i := 0; i < b.N; i++ {
		key := []byte{byte(i)}
		buf[i] = key
	}
	tree := New()
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		tree.Set(buf[i], buf[i])
		v, _ := tree.Get(buf[i])
		assert.Equal(b, v, buf[i])
	}
}
